feat: add static build script
